pyproject.toml
[tool.poetry]
name = "linkml-dh-templater"
version = "0.1.0"
description = ""
authors = ["Mark Andrew Miller <MAM@lbl.gov>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "*"
PyYAML = "*"
linkml = "1.1.12"
pandas = "^1.3.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
l2dh = "linkml_dh_templater.l2dh:l2dh_cli"
